[Modify] Dialog 컴포넌트 수ㅇ
- useDialogStore 생성
- 스토어 함수 변경 및 props 추가

# 161
